requirements.txt
numpy==1.22.3
opencv_python==4.6.0.66
pdf2image==1.16.0
Pillow==9.3.0
pytesseract==0.3.10
streamlit==1.14.0
tabula==1.0.5
tabula_py==2.5.1
